Moved _update_wrapped_in_try into abstract class. In that file, can check for tornado and skip class definition if not available
