+++ openmdao/visualization/realtime_plot/realtime_optimizer_plot.py
@@ -2,8 +2,6 @@
 
 from collections import defaultdict
 
-from tornado.ioloop import IOLoop
-
 from openmdao.utils.shell_proc import _is_process_running
 from openmdao.visualization.realtime_plot.realtime_plot_class import _RealTimePlot
 
@@ -292,14 +290,6 @@ class _RealTimeOptimizerPlot(_RealTimePlot):
 
         doc.add_periodic_callback(self._update_wrapped_in_try, callback_period)
         doc.title = "OpenMDAO Optimization Progress Plot"
-
-    def _update_wrapped_in_try(self):
-        try:
-            self._update()
-        except Exception as e:
-            print(f"Fatal error in periodic callback: {e}")
-            IOLoop.current().stop()
-            return
 
     def _update(self):
         # this is the main method of the class. It gets called periodically by Bokeh
